Add "getLedOffsetIndex" using custom addition
There was an overflow, when the number of LEDs is at least 64. In that
case the trail would skip a bunch because it used an signed 8 bit
integer. This adds a new function which calculcates the LED index after
adding two unsigned 8 bit integers. It would handle the overflow and
correctly apply modulo.