Fix snake animation
When placing an apple, it might determine, that the apple would be
placed inside the snake. In that case it'd add the length of the snake
itself (which slot isn't even to be guaranteed outside the snake), and
that could be outside the range. It would place the apple outside and
then the snake might never grow.

It also properly counts how many apples actually need to be placed.
Previously it didn't consider the already placed apples, so if the snake
only needed one apple and one already existed, it might still place one.

Lastly the direction is now random and not fixed to reverse.