Properly determine whether apple is below snake
When a new apple is spawned, the check whether it is below the snake was
flawed. Instead of trying again it just "continued" the iteration which
would only do the check.

Additionally added a method which returns the index relative to the
snake (e.g. = 0 is head, = _length - 1 is the tail). Also the body is
correctly stripped with a separate head.