Implementation using placement new for animations
Instead of using a stack allocated animation, it is now having a fixed
sized buffer. There it can allocate an animation using placement new,
without having to allocate anything new on the heap.

The advantage over stack allocating it, is that the loop itself
continues after the animation and thus can save state between
animations. This is not yet realized as it still calls the original loop
used with the stack allocated objects.