Enable extended unique ids and custom names
Instead of a fixed name, it can now be configured. Although it is not
really a secret, it is placed in the secrets.hpp as it should be defined
by the user.

It is also possible to use extended unique ids to support multiple
devices of them in the same home assistant instance.